⚡️ Add: 회원탈퇴시 프로필 이미지 삭제 및 예외 설정
생성된 그룹이 있거나 참가 그룹이 있으면 회원탈퇴 X
친구 관계가 있으면 회원탈퇴 X

탈퇴시 유저 프로필 이미지 삭제

+++ back/src/controllers/authController.js
@@ -27,6 +27,8 @@ const login = async (req, res, next) => {
     const id = req.user.id;
     const groups = await authService.findGroupsById(id);
     const friendships = await authService.findFriendIdsById(id);
+    const profileImage = await authService.findProfileImageById(id);
+    console.log(profileImage);
 
     const user = {
       id: id,
@@ -36,7 +38,6 @@ const login = async (req, res, next) => {
       groups: groups,
       friendshipsA: friendships,
     };
-    console.log(user);
     res.status(200).json(user);
   } catch (error) {
     console.error(error);
@@ -183,6 +184,20 @@ const changeInformation = async (req, res, next) => {
 const removeUser = async (req, res, next) => {
   try {
     const id = req.user.id;
+    const groups = await authService.findGroupsById(id);
+    const friendships = await authService.findFriendIdsById(id);
+    if (groups.length !== 0)
+      throw new Error('가입하거나 생성한 그룹이 있으면 탈퇴할 수 없습니다.');
+    if (friendships.length !== 0)
+      throw new Error('친구관계가 있으면 탈퇴할 수 없습니다.');
+
+    //프로필 이미지가 있g으면 있으면 삭제
+    await authService.deleteUserProfileImageById(id);
+
+    //개인 인증글 및 인증글의 댓글 삭제
+
+    //댓글의 자식(대댓글) 댓글 고유의 id 삭제
+
     const user = await authService.removeUser(id);
     console.log(user);
     res.status(200).json(user);
